# Meizon-Inventory-Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Inventory Management Service",
    "version": "1.0.1",
    "title": "Meizon Inventory Service",
    "license": {
      "name": "Apache Licence Version 2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/daily/readings/location/{locationId}": {
      "get": {
        "tags": [
          "daily-reading-controller-v-1"
        ],
        "summary": "getDailyReadingById",
        "operationId": "getDailyReadingByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "description": "endDate",
            "required": false,
            "type": "string"
          },
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "startDate",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "daily-reading-controller-v-1"
        ],
        "summary": "addNewDailyReading",
        "operationId": "addNewDailyReadingUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "dailyReadingDto",
            "description": "dailyReadingDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DailyReadingDto"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/daily/readings/user/{userId}": {
      "get": {
        "tags": [
          "daily-reading-controller-v-1"
        ],
        "summary": "getDailyReadingById",
        "operationId": "getDailyReadingByIdUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/daily/readings/{dailyReadingId}": {
      "get": {
        "tags": [
          "daily-reading-controller-v-1"
        ],
        "summary": "getDailyReadingById",
        "operationId": "getDailyReadingByIdUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "dailyReadingId",
            "in": "path",
            "description": "dailyReadingId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "daily-reading-controller-v-1"
        ],
        "summary": "deleteDailyReadingById",
        "operationId": "deleteDailyReadingByIdUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "dailyReadingId",
            "in": "path",
            "description": "dailyReadingId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "AddressDto": {
      "type": "object",
      "properties": {
        "street": {
          "type": "string"
        },
        "unitNumber": {
          "type": "string"
        }
      },
      "title": "AddressDto"
    },
...
    "DailyReadingDto": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "dailyReadingId": {
          "type": "string"
        },
        "dispensedVolume": {
          "type": "integer",
          "format": "int64"
        },
        "user": {
          "$ref": "#/definitions/UserDto"
        }
      },
      "title": "DailyReadingDto"
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/AddressDto"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "locationId": {
          "type": "integer",
          "format": "int64"
        },
        "mobile": {
          "type": "integer",
          "format": "int64"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserRoleDto"
          }
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "UserDto"
    },
    "UserPermissionDto": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "UserPermissionDto"
    },
    "UserRoleDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "permissions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserPermissionDto"
          }
        }
      },
      "title": "UserRoleDto"
    }
  }
}